Move role file from dreamrich to auth and rename auth to dr_auth
