+++ DreamRich/dr_auth/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class AuthConfig(AppConfig):
-    name = 'auth'
+    name = 'dr_auth'
