+++ DreamRich/employee/models.py
@@ -1,5 +1,5 @@
 from django.db import models
-from dreamrich import validators, roles
+from dreamrich import validators
 from client.models import ActiveClient
 from django.contrib.auth.models import User
 
@@ -17,6 +17,7 @@ class Employee(User):
         max_length=14,
         validators=[validators.validate_CPF]
     )
+
 
 class FinancialAdviser(Employee):
     clients = models.ManyToManyField(ActiveClient)
